Reorganizo archivos
    refs:#66
