+++ src/components/searchpage/filters/SeriesFilters.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import { ISerieApi } from '../../../api/SerieApi';
-import FilterSources from '../../common/filters/FilterSources';
-import FilterThemes from '../../common/filters/FilterThemes';
 import Filters from '../../style/Filters/Filters';
 import FilterSubTitle from '../../style/Filters/FilterSubTitle';
 import FilterTitle from '../../style/Filters/FilterTitle';
+import FilterSources from './FilterSources';
+import FilterThemes from './FilterThemes';
 
 
 interface IFilterProps {
